feat: add some validation and swagger

+++ src/controllers/productController.js
@@ -1,4 +1,5 @@
 const productService = require('../services/productService');
+const errMessages = require('../errors/errMessages');
 const error = require('../errors/appError');
 
 module.exports = {
@@ -14,6 +15,10 @@ module.exports = {
 
   async getById(req, res) {
     try {
+      if(!req.body.id) {
+        return res.status(400).json(new error({ status: 400, message: errMessages.BAD_DATA }));
+      }
+
       const product = await productService.getById(req.body.id);
       res.type('json');
       res.end(JSON.stringify(product));
@@ -24,6 +29,10 @@ module.exports = {
 
   async add(req, res) {
     try {
+      if(!(req.body || req.user.id)) {
+        return res.status(400).json(new error({ status: 400, message: errMessages.BAD_DATA }));
+      }
+
       const addedProduct = await productService.add(req.body, req.user.id);
       res.type('json');
       res.end(JSON.stringify(addedProduct));
@@ -34,6 +43,10 @@ module.exports = {
 
   async updateByUser(req, res) {
     try {
+      if(!(req.body || req.user)) {
+        return res.status(400).json(new error({ status: 400, message: errMessages.BAD_DATA }));
+      }
+
       const updatedProduct = await productService.updateByUser(req.body, req.user);
       res.type('json');
       res.end(JSON.stringify(updatedProduct));
@@ -44,6 +57,10 @@ module.exports = {
 
   async updateByAdmin(req, res) {
     try {
+      if(!req.body) {
+        return res.status(400).json(new error({ status: 400, message: errMessages.BAD_DATA }));
+      }
+
       const updatedProduct = await productService.updateByAdmin(req.body);
       res.type('json');
       res.end(JSON.stringify(updatedProduct));
@@ -54,6 +71,10 @@ module.exports = {
 
   async deleteByUser(req, res) {
     try {
+      if(!(req.body.id || req.user)) {
+        return res.status(400).json(new error({ status: 400, message: errMessages.BAD_DATA }));
+      }
+
       const deletedProduct = await productService.deleteByUser(req.body.id, req.user);
       res.type('json');
       res.end(JSON.stringify(deletedProduct));
@@ -64,7 +85,11 @@ module.exports = {
 
   async deleteByAdmin(req, res) {
     try {
-      const deletedProduct = await productService.deleteByAdmin(req.body.id);
+      if(!req.query.id) {
+        return res.status(400).json(new error({ status: 400, message: errMessages.BAD_DATA }));
+      }
+
+      const deletedProduct = await productService.deleteByAdmin(req.query.id);
       res.type('json');
       res.end(JSON.stringify(deletedProduct));
     } catch (ex) {
